refactor: extract spellSystem module from calculator (Phase 4a.3)
Move calculateGlobalMagicProfile and generateSpellsKnownSlots into
calculatorModules/spellSystem.ts. All snapshot tests pass with
identical output.
